Move makeotf tests that belong to makeotfexe
