+++ tests/makeotf_test.py
@@ -394,45 +394,6 @@ def test_build_options_cs_cl_bug459(args, input_filename, ttx_filename):
     assert differ([expected_ttx, actual_ttx, '-s', '<ttFont sfntVersion'])
 
 
-def test_bug416():
-    input_filename = "bug416/bug416.pfa"
-    input_featurename = "bug416/bug416.fea"
-    actual_path = _get_temp_file_path()
-    ttx_filename = "bug416.ttx"
-    runner(CMD + ['-n', '-o',
-                  'f', '_{}'.format(_get_input_path(input_filename)),
-                  'ff', '_{}'.format(_get_input_path(input_featurename)),
-                  'o', '_{}'.format(actual_path)])
-    actual_ttx = _generate_ttx_dump(actual_path, ['GPOS'])
-    expected_ttx = _get_expected_path(ttx_filename)
-    assert differ([expected_ttx, actual_ttx, '-s', '<ttFont sfntVersion'])
-
-
-def test_bug438():
-    """ The feature file bug438/feat.fea contains languagesystem
-    statements for a language other than 'dflt' with the 'DFLT' script
-    tag. With the fix, makeotf will build an otf which is identical to
-    't1pfa-dev.ttx'. Without the fix, it will fail to build an OTF."""
-    input_filename = T1PFA_NAME
-    feat_filename = 'bug438/feat.fea'
-    ttx_filename = 't1pfa-dev.ttx'
-    actual_path = _get_temp_file_path()
-    runner(CMD + ['-n', '-o',
-                  'f', '_{}'.format(_get_input_path(input_filename)),
-                  'ff', '_{}'.format(_get_input_path(feat_filename)),
-                  'o', '_{}'.format(actual_path)])
-    actual_ttx = _generate_ttx_dump(actual_path)
-    expected_ttx = _get_expected_path(ttx_filename)
-    assert differ([expected_ttx, actual_ttx,
-                   '-s',
-                   '<ttFont sfntVersion' + SPLIT_MARKER +
-                   '    <checkSumAdjustment value=' + SPLIT_MARKER +
-                   '    <checkSumAdjustment value=' + SPLIT_MARKER +
-                   '    <created value=' + SPLIT_MARKER +
-                   '    <modified value=',
-                   '-r', r'^\s+Version.*;hotconv.*;makeotfexe'])
-
-
 @pytest.mark.parametrize('args, font, fontinfo', [
     (['cn'], 'type1', ''),
     (['cn'], 'no_notdef', ''),
@@ -521,17 +482,3 @@ def test_GOADB_options_bug497(opts):
                    '    <created value=' + SPLIT_MARKER +
                    '    <modified value=',
                    '-r', r'^\s+Version.*;hotconv.*;makeotfexe'])
-
-
-def test_useMarkFilteringSet_flag_bug196():
-    input_filename = "bug196/bug196.pfa"
-    input_featurename = "bug196/bug196.fea"
-    actual_path = _get_temp_file_path()
-    ttx_filename = "bug196.ttx"
-    runner(CMD + ['-n', '-o',
-                  'f', '_{}'.format(_get_input_path(input_filename)),
-                  'ff', '_{}'.format(_get_input_path(input_featurename)),
-                  'o', '_{}'.format(actual_path)])
-    actual_ttx = _generate_ttx_dump(actual_path, ['GSUB'])
-    expected_ttx = _get_expected_path(ttx_filename)
-    assert differ([expected_ttx, actual_ttx, '-s', '<ttFont sfntVersion'])
